Move get_themes to pydoctor.themes and lower the treshold to use importlib_resources compatibility package to python3.7 when possible.
