+++ pydoctor/driver.py
@@ -9,6 +9,7 @@ import sys
 import warnings
 from inspect import getmodulename
 
+from pydoctor.themes import get_themes
 from pydoctor import model, zopeinterface, __version__
 from pydoctor.templatewriter import IWriter, TemplateError, TemplateLookup
 from pydoctor.sphinx import (MAX_AGE_HELP, USER_INTERSPHINX_CACHE,
@@ -19,8 +20,9 @@ if TYPE_CHECKING:
 else:
     NoReturn = None
 
-# Newer APIs from importlib_resources should arrive to stdlib importlib.resources in Python 3.9.
-if sys.version_info < (3, 9):
+# On Python 3.7+, use importlib.resources from the standard library.
+# On older versions, a compatibility package must be installed from PyPI.
+if sys.version_info < (3, 7):
     import importlib_resources
 else:
     import importlib.resources as importlib_resources
@@ -93,15 +95,6 @@ def get_supported_docformats() -> Iterator[str]:
             continue
         else:
             yield moduleName
-
-def get_themes() -> Iterator[str]:
-    """
-    Get the list of the available themes.
-    """
-    for name in importlib_resources.contents('pydoctor.themes'):
-        if (not name.startswith('_') and not
-            importlib_resources.is_resource('pydoctor.themes', name)) :
-            yield name
 
 class CustomOption(Option):
     TYPES = Option.TYPES + ("path",)
